refactor: move message utils into message.ts
