+++ packages/messaging/src/index.ts
@@ -17,4 +17,3 @@
  */
 
 export * from './Message.js'
-export * from './Message.utils.js'
